zigma_backend/requirements.txt
WINDOWS:
Django==2.1.8
django-filter==2.1.0
django-pyodbc-azure==2.1.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-queryfields==1.0.0
django-filter==2.1.0
Pillow==6.0.0
pyodbc==4.0.26
pytz==2019.1
sqlparse==0.3.0

MAC:
Markdown==3.1
stevedore==1.30.1


----Osäker på om dessa behövs----
python-dateutil==2.8.0
python-mimeparse==1.6.0
pbr==5.1.3
six==1.12.0
